backend/requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
environs==11.0.0
Faker==27.0.0
marshmallow==3.21.3
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
djoser==2.2.3
